requirements.txt
python-dotenv
streamlit
langchain
langchain-community
sentence-transformers
faiss-cpu
huggingface_hub
pypdf
transformers
torch
numpy

